III-3113 Move ContextDecoratedCommandBus
