+++ app/CommandHandling/CommandBusServiceProvider.php
@@ -18,7 +18,6 @@ use CultuurNet\UDB3\Role\ValueObjects\Permission;
 use CultuurNet\UDB3\Security\ClassNameCommandFilter;
 use CultuurNet\UDB3\Security\CultureFeedUserIdentification;
 use CultuurNet\UDB3\Security\SecurityWithUserPermission;
-use CultuurNet\UDB3\Silex\ContextDecoratedCommandBus;
 use CultuurNet\UDB3\Silex\Labels\LabelServiceProvider;
 use Qandidate\Toggle\ToggleManager;
 use Silex\Application;
